move Unmarshal/Marshal to the Message
