+++ field/composite_test.go
@@ -251,15 +251,14 @@ func TestTLVPacking(t *testing.T) {
 	})
 
 	t.Run("Unpack correctly deserialises bytes to the data struct", func(t *testing.T) {
-		data := &TLVTestData{}
-
 		composite := NewComposite(tlvTestSpec)
-		err := composite.SetData(data)
-		require.NoError(t, err)
 
 		read, err := composite.Unpack([]byte{0x30, 0x31, 0x34, 0x9a, 0x3, 0x21, 0x7, 0x20, 0x9f, 0x2, 0x6, 0x0, 0x0, 0x0, 0x0, 0x5, 0x1})
 		require.NoError(t, err)
 		require.Equal(t, 17, read)
+
+		data := &TLVTestData{}
+		require.NoError(t, composite.UnmarshalValue(data))
 
 		require.Equal(t, "210720", data.F9A.Value)
 		require.Equal(t, "000000000501", data.F9F02.Value)
@@ -361,13 +360,14 @@ func TestCompositePacking(t *testing.T) {
 			F1 *Numeric
 		}
 		composite := NewComposite(compositeTestSpec)
-		err := composite.SetData(&TestDataIncorrectType{})
-		require.NoError(t, err)
 
-		read, err := composite.Unpack([]byte("ABCD12"))
-		require.Equal(t, 0, read)
+		_, err := composite.Unpack([]byte("ABCD12"))
+
+		data := &TestDataIncorrectType{}
+		err = composite.UnmarshalValue(data)
+
 		require.Error(t, err)
-		require.EqualError(t, err, "failed to set data for field 1: data does not match required *String type")
+		require.EqualError(t, err, "failed to get data from field 1: data does not match required *String type")
 	})
 
 	t.Run("Unpack returns an error on failure of subfield to unpack bytes", func(t *testing.T) {
@@ -463,15 +463,14 @@ func TestCompositePacking(t *testing.T) {
 	})
 
 	t.Run("Unpack correctly deserialises bytes to the data struct", func(t *testing.T) {
-		data := &CompositeTestData{}
-
 		composite := NewComposite(compositeTestSpec)
-		err := composite.SetData(data)
-		require.NoError(t, err)
 
 		read, err := composite.Unpack([]byte("ABCD12"))
 		require.Equal(t, compositeTestSpec.Length, read)
 		require.NoError(t, err)
+
+		data := &CompositeTestData{}
+		require.NoError(t, composite.UnmarshalValue(data))
 
 		require.Equal(t, "AB", data.F1.Value)
 		require.Equal(t, "CD", data.F2.Value)
@@ -480,14 +479,13 @@ func TestCompositePacking(t *testing.T) {
 	})
 
 	t.Run("SetBytes correctly deserialises bytes to the data struct", func(t *testing.T) {
-		data := &CompositeTestData{}
-
 		composite := NewComposite(compositeTestSpec)
-		err := composite.SetData(data)
+
+		err := composite.SetBytes([]byte("ABCD12"))
 		require.NoError(t, err)
 
-		err = composite.SetBytes([]byte("ABCD12"))
-		require.NoError(t, err)
+		data := &CompositeTestData{}
+		require.NoError(t, composite.UnmarshalValue(data))
 
 		require.Equal(t, "AB", data.F1.Value)
 		require.Equal(t, "CD", data.F2.Value)
@@ -773,16 +771,15 @@ func TestCompositePackingWithTags(t *testing.T) {
 	})
 
 	t.Run("Unpack correctly deserialises out of order composite subfields to the data struct", func(t *testing.T) {
-		data := &CompositeTestData{}
-
 		composite := NewComposite(compositeTestSpecWithTagPadding)
-		err := composite.SetData(data)
-		require.NoError(t, err)
 
 		read, err := composite.Unpack([]byte("280202CD0302120102AB11060102YZ"))
 
 		require.NoError(t, err)
 		require.Equal(t, 30, read)
+
+		data := &CompositeTestData{}
+		require.NoError(t, composite.UnmarshalValue(data))
 
 		require.Equal(t, "AB", data.F1.Value)
 		require.Equal(t, "CD", data.F2.Value)
@@ -791,13 +788,13 @@ func TestCompositePackingWithTags(t *testing.T) {
 	})
 
 	t.Run("Unpack correctly deserialises out of order composite subfields to the unpadded data struct", func(t *testing.T) {
-		data := &CompositeTestDataWithoutTagPadding{}
-
 		composite := NewComposite(compositeTestSpecWithoutTagPadding)
-		err := composite.SetData(data)
-		require.NoError(t, err)
 
 		read, err := composite.Unpack([]byte("120202CD0102AB"))
+
+		data := &CompositeTestDataWithoutTagPadding{}
+
+		require.NoError(t, composite.UnmarshalValue(data))
 
 		require.NoError(t, err)
 		require.Equal(t, 14, read)
@@ -807,16 +804,15 @@ func TestCompositePackingWithTags(t *testing.T) {
 	})
 
 	t.Run("Unpack correctly deserialises partial subfields to the data struct", func(t *testing.T) {
-		data := &CompositeTestData{}
-
 		composite := NewComposite(compositeTestSpecWithTagPadding)
-		err := composite.SetData(data)
-		require.NoError(t, err)
 
 		read, err := composite.Unpack([]byte("120302120102AB"))
 
 		require.NoError(t, err)
 		require.Equal(t, 14, read)
+
+		data := &CompositeTestData{}
+		require.NoError(t, composite.UnmarshalValue(data))
 
 		require.Equal(t, "AB", data.F1.Value)
 		require.Nil(t, data.F2)
@@ -824,11 +820,7 @@ func TestCompositePackingWithTags(t *testing.T) {
 	})
 
 	t.Run("Unpack correctly ignores excess bytes in excess of the length described by the prefix", func(t *testing.T) {
-		data := &CompositeTestData{}
-
 		composite := NewComposite(compositeTestSpecWithTagPadding)
-		err := composite.SetData(data)
-		require.NoError(t, err)
 
 		// "11060102YZ" falls outside of the bounds of the 18 byte limit imposed
 		// by the prefix. Therefore, F11 must be nil.
@@ -836,6 +828,9 @@ func TestCompositePackingWithTags(t *testing.T) {
 
 		require.NoError(t, err)
 		require.Equal(t, 20, read)
+
+		data := &CompositeTestData{}
+		require.NoError(t, composite.UnmarshalValue(data))
 
 		require.Equal(t, "AB", data.F1.Value)
 		require.Equal(t, "CD", data.F2.Value)
