+++ message.go
@@ -2,8 +2,10 @@ package iso8583
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"reflect"
+	"regexp"
 	"sort"
 	"strconv"
 
@@ -127,6 +129,11 @@ func (m *Message) GetBytes(id int) ([]byte, error) {
 
 func (m *Message) GetField(id int) field.Field {
 	return m.fields[id]
+}
+
+func (m *Message) IsFieldValueSet(id int) bool {
+	_, ok := m.fieldsMap[id]
+	return ok
 }
 
 // Fields returns the map of the set fields
@@ -377,4 +384,137 @@ func (m *Message) Clone() (*Message, error) {
 	}
 
 	return newMessage, nil
+}
+
+// TODO: Marshal traverses v and sets message field value into the corresponding
+// struct field value pointed by v. If v is nil or not a pointer it returns
+// error.
+func (m *Message) Marshal(v interface{}) error {
+	rv := reflect.ValueOf(v)
+	if rv.Kind() != reflect.Ptr || rv.IsNil() {
+		return errors.New("data is not a pointer or nil")
+	}
+
+	// get the struct from the pointer
+	dataStruct := rv.Elem()
+
+	if dataStruct.Kind() != reflect.Struct {
+		return errors.New("data is not a struct")
+	}
+
+	// iterate over struct fields
+	for i := 0; i < dataStruct.NumField(); i++ {
+		fieldIndex, err := getFieldIndex(dataStruct.Type().Field(i))
+		if err != nil {
+			return fmt.Errorf("getting field %d index: %w", i, err)
+		}
+
+		// skip field without index
+		if fieldIndex < 0 {
+			continue
+		}
+
+		messageField := m.GetField(fieldIndex)
+		// if struct field we are usgin to populate value expects to
+		// set index of the field that is not described by spec
+		if messageField == nil {
+			return fmt.Errorf("no message field defined by spec with index: %d", fieldIndex)
+		}
+
+		dataField := dataStruct.Field(i)
+		if dataField.IsNil() {
+			continue
+		}
+
+		err = messageField.MarshalValue(dataField.Interface())
+		if err != nil {
+			return fmt.Errorf("failed to set value to field %d: %w", fieldIndex, err)
+		}
+
+		m.fieldsMap[fieldIndex] = struct{}{}
+	}
+
+	return nil
+}
+
+// Unmarshal traverses the message fields recursively and for each field, sets
+// the field value into the corresponding struct field value pointed by v. If v
+// is nil or not a pointer it returns error.
+func (m *Message) Unmarshal(v interface{}) error {
+	rv := reflect.ValueOf(v)
+	if rv.Kind() != reflect.Ptr || rv.IsNil() {
+		return errors.New("data is not a pointer or nil")
+	}
+
+	// get the struct from the pointer
+	dataStruct := rv.Elem()
+
+	if dataStruct.Kind() != reflect.Struct {
+		return errors.New("data is not a struct")
+	}
+
+	// iterate over struct fields
+	for i := 0; i < dataStruct.NumField(); i++ {
+		fieldIndex, err := getFieldIndex(dataStruct.Type().Field(i))
+		if err != nil {
+			return fmt.Errorf("getting field %d index: %w", i, err)
+		}
+
+		// skip field without index
+		if fieldIndex < 0 {
+			continue
+		}
+
+		// we can get data only if field value is set
+		messageField := m.GetField(fieldIndex)
+		if messageField == nil {
+			continue
+		}
+
+		if _, set := m.fieldsMap[fieldIndex]; !set {
+			continue
+		}
+
+		dataField := dataStruct.Field(i)
+		if dataField.IsNil() {
+			dataField.Set(reflect.New(dataField.Type().Elem()))
+		}
+
+		err = messageField.UnmarshalValue(dataField.Interface())
+		if err != nil {
+			return fmt.Errorf("failed to get value from field %d: %w", fieldIndex, err)
+		}
+	}
+
+	return nil
+}
+
+var fieldNameIndexRe = regexp.MustCompile(`^F\d+$`)
+
+// fieldIndex returns index of the field. First, it checks field name. If it
+// does not match FNN (when NN is digits), it checks value of `index` tag.  If
+// negative value returned (-1) then index was not found for the field.
+func getFieldIndex(field reflect.StructField) (int, error) {
+	dataFieldName := field.Name
+
+	if len(dataFieldName) > 0 && fieldNameIndexRe.MatchString(dataFieldName) {
+		indexStr := dataFieldName[1:]
+		fieldIndex, err := strconv.Atoi(indexStr)
+		if err != nil {
+			return -1, fmt.Errorf("converting field index into int: %w", err)
+		}
+
+		return fieldIndex, nil
+	}
+
+	if indexStr := field.Tag.Get("index"); indexStr != "" {
+		fieldIndex, err := strconv.Atoi(indexStr)
+		if err != nil {
+			return -1, fmt.Errorf("converting field index into int: %w", err)
+		}
+
+		return fieldIndex, nil
+	}
+
+	return -1, nil
 }
